Reorganize the source code
Move all the code to pysrc - but leave the tests where they are. This
patch also updates the python path in run_test.sh and updates the
clean.sh script to recursively delete the annoying files. The tests.sh
script has been deleted as it was useless.

Maybe in the future such a script could be useful but for now tests are
just run by hand via the run_test.sh script.

Signed-off-by: Alex Waterman <alexw@nvidia.com>