Add notion of a 'lot'
Add the ability to track lots of stocks. Now, when adding sells
to a portfolio, a cost basis can be computed. The cost basis
computation assumes that investors always sell the lot with
the lowest tax penalty (i.e lowest gain) first.

Of course there's other strategies but this seems like a sensible
approach.

Signed-off-by: Alex Waterman <alexw@nvidia.com>